generator: Support Extent64 arg in Go code
See the earlier commit "Add Extent64 arg type" for rationale in
supporting a new generator arg type.  This patch adds the Go bindings
for use of Extent64, which required adding a counterpart Go type and
introducing a helper function for copying from the C array to an
actual Go object, very similar to the existing copy_uint32_array
helper for 32-bit extents.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-ID: <20230803015720.1090260-7-eblake@redhat.com>
Reviewed-by: Richard W.M. Jones <rjones@redhat.com>