generator: Support Extent64 arg in OCaml code
See the earlier commit "Add Extent64 arg type" for rationale in
supporting a new generator arg type.  This patch adds the OCaml
bindings for use of Extent64, which required adding a counterpart
OCaml type.

Note that we intend to guarantee that the size is always a positive
value.  If we were using using structured_reply_in_bounds() like we do
for REPLY_TYPE_DATA, this would be trivial; but in practice, the NBD
protocol says we have to accept server responses that extend beyond
the bounds of the client's request.  Still, the state machine can
easily check that the server's cumulative size has not exceeded the
export's length, which in practice is limited by off_t; furthermore,
the NBD protocol allows truncation of the response as long as the
client makes progress, and the client does not have to know whether
that truncation happened at the server or in our state machine.
Therefore, our assertion of a positive size is viable.

However, the flags can be negative: OCaml's only native 64-bit integer
is inherently signed, but at least it does have a logical shift
operator for performing unsigned bit-twiddling operations.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-ID: <20230803015720.1090260-8-eblake@redhat.com>
Reviewed-by: Richard W.M. Jones <rjones@redhat.com>