info: Update nbdinfo --map to use 64-bit block status
Although we usually map "base:allocation" which doesn't require the
use of the 64-bit API for flags, this application IS intended to map
out other metacontexts that might have 64-bit flags.  And when
extended headers are in use, we might as well ask for the server to
give us extents as large as it wants, rather than breaking things up
at 4G boundaries.

At the time this patch was written, there are no known servers that
actually provide a metacontext with 64-bit flags.  However, that is
planned for the nbdkit v3 protocol.

Signed-off-by: Eric Blake <eblake@redhat.com>
Reviewed-by: Richard W.M. Jones <rjones@redhat.com>
Message-ID: <20230803015720.1090260-18-eblake@redhat.com>