generator: indent C argument list 2 spaces relative to function designator
Change the optional "parens" parameter of "print_arg_list" from bool to a
new union type. The new union type has three data constructors, NoParens
(matching the previous "false" value), ParensSameLine (matching the
previous "true" value), and a third constructor called
"ParensNewLineWithIndent", which wraps an "int".

[ParensNewLineWithIndent n] stands for the following formatting construct
(with wrapping):

> ret_type foobar (
>            type1 param1, type2 param2, type3 param3, type4 param4,
>            type5 param5
>          );
> <---n--->

and without wrapping:

> ret_type foobar (
>            type1 param1, type2 param2, type3 param3, type4 param4, type5 param5
>          );
> <---n--->

Capture "n" (that is, the 0-based column where the function designator
starts) in "print_call", and put the new formatting to use.

In effect, still utilize "pr_wrap", for wrapping the argument list at
comma (",") characters, but shift the left hand side of the wrapped block
from the end of the function designator near the front of it.

The use case is: if "foobar" is long, and at least one "typeN" is also
long, previously we would get

> ret_type extremely_long_verbose_foobar (extremely_long_verbose_type1 param1,
>                                         type2 param2, type3 param3,
>                                         type4 param4, type5 param5);

That is, the first wrapping opportunity on a particular line would be too
late already, making the source code too wide.

This patch changes the generated files "include/libnbd.h" and
"lib/unlocked.h". Taking the declaration responsible for the longest line
in each of those header files, the declaration is reformatted as shown
below:

[include/libnbd.h]

> -extern int nbd_aio_opt_list_meta_context_queries (struct nbd_handle *h,
> -                                                  char **queries,
> -                                                  nbd_context_callback context_callback,
> -                                                  nbd_completion_callback completion_callback)
> +extern int nbd_aio_opt_list_meta_context_queries (
> +             struct nbd_handle *h, char **queries,
> +             nbd_context_callback context_callback,
> +             nbd_completion_callback completion_callback
> +           )
>      LIBNBD_ATTRIBUTE_NONNULL (1, 2);

[lib/unlocked.h]

> -extern int nbd_unlocked_aio_opt_list_meta_context_queries (struct nbd_handle *h,
> -                                                           char **queries,
> -                                                           nbd_context_callback *context_callback,
> -                                                           nbd_completion_callback *completion_callback)
> +extern int nbd_unlocked_aio_opt_list_meta_context_queries (
> +             struct nbd_handle *h, char **queries,
> +             nbd_context_callback *context_callback,
> +             nbd_completion_callback *completion_callback
> +           )
>      LIBNBD_ATTRIBUTE_NONNULL (1, 2);

Update the GoLang generator to stick with "no parens".

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2172516
Signed-off-by: Laszlo Ersek <lersek@redhat.com>
Message-Id: <20230421073418.146910-5-lersek@redhat.com>
Reviewed-by: Richard W.M. Jones <rjones@redhat.com>