generator/C: print_trace_enter: expand "function" shorthand in value iters
In "print_trace_enter", when we iterate over the "args" and "optargs"
lists with "List.iter" for printing the *values*, we use the "function"
shorthand within the functions that get applied to each arg / optarg
element. Expand the "function" shorthand to "fun" + "match", so we can
more easily modify the code subsequently.

This patch is purely refactoring, it does not change the generated output.
It is best viewed with "git show -b".

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2172516
Signed-off-by: Laszlo Ersek <lersek@redhat.com>
Message-Id: <20230425071108.82663-13-lersek@redhat.com>
Reviewed-by: Eric Blake <eblake@rehat.com>
Reviewed-by: Richard W.M. Jones <rjones@redhat.com>