generator/C: print_trace_enter: generate commas as suffixes, not prefixes
The iteration creating the "args" values and the iteration creating the
"optargs" values both print commas as prefixes; in other words, whenever
any pattern matches, the "pr" that's invoked starts with a ", " substring.

For the sake of a subsequent patch, print commas as suffixes, not as
prefixes:

- Once we're done outputting the format string, calculate the total number
  of arguments (args + optargs).

- If there is at least one argument to print, follow the format string
  with a comma-space (", ").

- Turn both iterations (i.e., over "args" and "optargs") into numbered
  iterations (i.e., use "List.iteri").

- From each pattern's own "pr" call, strip the comma-space (", ") prefix.

- After the "match", see if the (opt)arg just handled was the last one
  against the *total* number of arguments; if not, print the comma-space
  as a suffix.

This patch is purely refactoring, it does not change the generated output.

The command "git show --color-words" is somewhat helpful for viewing the
patch.

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2172516
Signed-off-by: Laszlo Ersek <lersek@redhat.com>
Message-Id: <20230425071108.82663-15-lersek@redhat.com>
Reviewed-by: Eric Blake <eblake@rehat.com>
Reviewed-by: Richard W.M. Jones <rjones@redhat.com>