generator/C: print_trace_leave: separate arg list from function designator
Dependent on "may_set_error", we (a) call debug() vs. debug_direct(), (b)
don't pass vs. pass the API name to the function from (a).

Currently this is (justifiedly) handled within a single "if"; however, for
wrapping the argument lists later -- with differrent indentations --,
let's now split the arg lists to separate invocations of "pr", at the cost
of another "if may_set_error".

This is a refactoring; it does not change the generated output.

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2172516
Signed-off-by: Laszlo Ersek <lersek@redhat.com>
Message-Id: <20230425071108.82663-20-lersek@redhat.com>
Reviewed-by: Eric Blake <eblake@rehat.com>
Reviewed-by: Richard W.M. Jones <rjones@redhat.com>