generator/C: print_flags_check: break guard condition to a new line
Break the flags guard condition, if any, to a new line.

One of the most impactful changes of this patch is [lib/api.c]:

> @@ -4218,7 +4224,8 @@ nbd_zero (struct nbd_handle *h, uint64_t
>      ret = -1;
>      goto out;
>    }
> -  if (unlikely ((flags & ~0x13) != 0) && ((h->strict & LIBNBD_STRICT_FLAGS) || flags > UINT16_MAX)) {
> +  if (unlikely ((flags & ~0x13) != 0) &&
> +      ((h->strict & LIBNBD_STRICT_FLAGS) || flags > UINT16_MAX)) {
>      set_error (EINVAL, "%s: invalid value for flag: 0x%x",
>                 "flags", flags);
>      ret = -1;

Note that this patch is not really future-proof; if we add another

  guard = Some "...";

to "generator/API.ml", that string can still exceed 80 characters.

For now, the current patch should suffice.

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2172516
Signed-off-by: Laszlo Ersek <lersek@redhat.com>
Message-Id: <20230425071108.82663-22-lersek@redhat.com>
Reviewed-by: Eric Blake <eblake@rehat.com>
Reviewed-by: Richard W.M. Jones <rjones@redhat.com>