internal: Align packed members in union sbuf
While analyzing 'union sbuf' in preparation to add more members to the
union, I noticed several things related to __attribute__((packed))
that can be improved.  It helps to note that that the bulk of the
members of 'union sbuf' are already marked as packed structures from
nbd-protocol.h, where we don't need to repeat that annotation in
internal.h but where it does factor into sbuf alignment.

In particular, even when using a packed struct, there are benefits to
aligning the packed struct to natural alignment insofar as possible
(hardware that supports unaligned reads still performs better on
natural alignments, and a compiler that has to inject code to avoid
SIGBUS for unaligned access on picky hardware can elide that code when
it can prove a particular region of memory is naturally aligned
despite being packed).

Pre-patch, on x86_64, we have:

struct nbd_handle {
  ...
  size_t wlen;          /* 8-byte aligned, 8 bytes, offset % 8 == 0 */
  int wflags;           /* 4-byte aligned, 4 bytes, offset % 8 == 0 */
  /* no padding needed */
  union {
    ...
    struct {
      struct nbd_structured_reply structured_reply;
                        /* 1-byte aligned, 20 bytes, offset % 8 == 4 */
      /* no padding needed */
      union {
        struct nbd_structured_reply_offset_data offset_data;
                        /* 1-byte aligned, 8 bytes, offset % 8 == 0 */
        ...
      } payload;        /* 1-byte aligned, M bytes, offset % 8 == 0 */
    } sr;               /* 1-byte aligned, M+20 bytes, offset % 8 == 4 */
    ... all other members are either packed or at most uint32_t ...
  } sbuf;               /* 4-byte aligned, N bytes, offset % 8 == 4 */
  ...
};

That is, the compiler _happens_ to have aligned access for
sbuf.sr.payload.offset_data.offset, but only by accident, while the
uint64_t in sbuf.sr.structured_reply is unaligned.  Yet both of those
uint64_t members happen to have natural alignment if their respective
packed struct is naturally aligned, and accesses to sbuf.sr appear on
the hot path.  So we can do better by telling the compiler it is okay
to insert padding where it is useful, via the judicious use of of a
non-packed uint64_t inside any union that would otherwise not pick up
minimum alignments from its packed members.  Post-patch, we now have
on x86_64:

struct nbd_handle {
  ...
  size_t wlen;          /* 8-byte aligned, 8 bytes, offset % 8 == 0 */
  int wflags;           /* 4-byte aligned, 4 bytes, offset % 8 == 0 */
  char __pad[4];        /* 1-byte aligned, 4 bytes, offset % 8 == 4 */
  union {
    uint64_t align_;    /* 8-byte aligned, 8 bytes, offset % 8 == 0 */
    ...
    struct {
      struct nbd_structured_reply structured_reply;
                        /* 1-byte aligned, 20 bytes, offset % 8 == 0 */
      char __pad[4];    /* 1-byte aligned, 4 bytes, offset % 8 == 4 */
      union {
        uint64_t align_;/* 8-byte aligned, 8 bytes, offset % 8 == 0 */
        struct nbd_structured_reply_offset_data offset_data;
                        /* 1-byte aligned, 8 bytes, offset % 8 == 0 */
        ...
      } payload;        /* 8-byte aligned, M bytes, offset % 8 == 0 */
    } sr;               /* 8-byte aligned, M+24 bytes, offset % 8 == 0 */
    ... all other members are either packed or at most uint32_t ...
  } sbuf;               /* 8-byte aligned, N bytes, offset % 8 == 0 */
  ...
};

which gets natural alignment for both sbuf.sr.structured_reply.cookie
and sbuf.sr.payload.offset_data.data.  Note that sbuf.or does not
benefit from this treatment (currently,
sbuf.or.payload.export.exportsize is the only uint64_t in that
particular payload union, but does not occur at a natural alignment to
begin with); but it is also worth remembering that option negotiation
is not in the hot path the way sbuf.sr is.

Actually testing the alignment change from adding align_ members is
harder to do, especially without C11's <stdalign.h>, in part because
on 32-bit platforms where uint64_t is only 4-byte aligned (at least,
when compiling without gcc -malign-double), there is no change to
layout (my choice to overlay an unused uint64_t into unions, rather
than allocating explicit padding space before the union, was
intentional).

Not touched here: gcc's -Wpacked says several (but not all) of our
structures in nbd-protocol.h are already naturally aligned, where
adding a packed notation can actually cause slower code when embedding
that type in a larger structure (exactly what we're doing when
combining those structs into sbuf), when the compiler may have to
prepare for unaligned access even if the struct would otherwise be
aligned.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20230531021013.1604692-1-eblake@redhat.com>
[eblake: split into 3 patches, per Laszlo's recommendation]
Reviewed-by: Laszlo Ersek <lersek@redhat.com>
Reviewed-by: Richard W.M. Jones <rjones@redhat.com>