internal: Drop useless attribute packed in union sbuf
While analyzing 'union sbuf' in preparation to add more members to the
union, I noticed several things related to __attribute__((packed))
that can be improved.  For this patch, it helps to note that that the
bulk of the members of 'union sbuf' are already marked as packed
structures from nbd-protocol.h, where we don't need to repeat that
annotation in internal.h.

Two of the union members are themselves sub-structs with two parts:
both sbuf.or and sbuf.sr need distinct storage for a header (a packed
structure) and an optional payload (a union of various items;
currently each member of both of those unions are already packed),
where the choice of union branch and overall size of the payload to be
read (if any) is determined by information in the header.  Later
states then refer to information from both the header and payload, so
we need to keep the sub-structs (rather than hoisting the two parts of
the sub-struct into being directly-overlapping top-level members of
union sbuf proper).  But because we don't know the payload size until
after the header is read, the state machine uses separate recv() calls
into the two parts of sbuf.or and sbuf.sr; and there is no specific
reason that the two reads need to occur into adjacent memory.  Thus,
these two packed annotations buy us nothing at this layer and can be
safely elided.

Note that this patch in isolation makes NO difference to the struct
layout: gcc treats any union whose members are all packed as having a
natural alignment of 1, so even with the attribute on sbuf.or and
sbuf.sr removed, sbuf.or.payload and sbuf.sr.payload are still placed
immediately after the previous member.  But a later patch plans to add
a non-packed uint64_t to sbuf.sr.payload for alignment reasons, and
that addition has no effect if the packed attribute removal in this
patch is reverted.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20230531021013.1604692-1-eblake@redhat.com>
[eblake: split into 3 patches, per Laszlo's recommendation]
Reviewed-by: Laszlo Ersek <lersek@redhat.com>
Reviewed-by: Richard W.M. Jones <rjones@redhat.com>