generator: Rename states-reply-structured to states-reply-chunk
Upcoming patches to add extended headers want to reuse most of the
common payload parser with structured replies (only the block status
reply payload differs between structured and extended, and even that
difference can share quite a bit of low-level code).  Renaming the
file and the associated states and structs from "structured" to
"chunk" makes it more obvious that we will be sharing the code
independent from the header style parsed in the earlier REPLY portion
of the state machine.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20230609021738.3151009-4-eblake@redhat.com>
[eblake: squash in struct renames to match file rename]
Acked-by: Laszlo Ersek <lersek@redhat.com>