states: Document our reliance on type overlaps
Back in 2016 when I was helping with the effort to add structured
replies to the NBD spec [1], we intentionally chose a wire layout
where the magic number and cookie overlap, even while the middle
member changes from uint32_t error to the pair uint16_t flags and
type.  Based only on a strict reading of C rules on effective types
and compatible type prefixes, it's probably questionable on whether my
reliance on type aliasing to reuse cookie from the same offset of a
union, or even the fact that a structured reply is built by first
reading bytes into sbuf.simple_reply then following up with only bytes
into the tail of sbuf.sr.structured_reply is strictly portable.  But
since it works in practice, it's worth at least adding some compile-
and run-time assertions that our (ab)use of aliasing is accessing the
bytes we want under the types we expect.  Upcoming patches will
restructure part of the sbuf layout to hopefully be a little easier to
tie back to strict C standards.

[1] https://github.com/NetworkBlockDevice/nbd/commit/7ff2e45e4

Suggested-by: Laszlo Ersek <lersek@redhat.com>
Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20230609021738.3151009-2-eblake@redhat.com>
Reviewed-by: Laszlo Ersek <lersek@redhat.com>
[eblake: also test for equal sizeof relation, add nbd.git commit id]